remove my own training.
